Moved SubView tests to own file
